books-ai-publishing/requirements.txt
# 三书AI注疏出版系统 - Python依赖

# Core AI SDK
anthropic>=0.39.0          # Claude API (Anthropic官方SDK)

# File processing
pyyaml>=6.0                # YAML配置文件解析
python-frontmatter>=1.0.0  # Markdown front matter处理

# Text processing
markdown>=3.5.0            # Markdown解析和生成
beautifulsoup4>=4.12.0     # HTML清洗（如果需要网页抓取）
lxml>=5.0.0                # XML/HTML解析

# Document generation
pandoc>=2.3                # Markdown转PDF/EPUB（Python wrapper）
pypandoc>=1.11             # Pandoc Python接口

# PDF generation
reportlab>=4.0.0           # PDF生成（备选方案）
weasyprint>=60.0           # HTML/CSS转PDF

# Web scraping (用于Agent 2检索资料)
requests>=2.31.0           # HTTP请求
beautifulsoup4>=4.12.0     # HTML解析

# Utilities
python-dotenv>=1.0.0       # 环境变量管理
click>=8.1.0               # CLI工具
rich>=13.0.0               # 漂亮的终端输出
tqdm>=4.66.0               # 进度条

# Optional: LangChain (如果需要高级prompt工程)
# langchain>=0.1.0
# langchain-anthropic>=0.1.0

# Development
pytest>=7.4.0              # 测试框架
black>=23.0.0              # 代码格式化
flake8>=6.1.0              # 代码检查
mypy>=1.7.0                # 类型检查
